RAG-pdf/pyproject.toml
[project]
name = "rag-notes"
version = "0.1.0"
description = ""
authors = [
    { name = "Priyanshu-K07", email = "priyanshukapadnis07@gmail.com" }
]
requires-python = ">=3.12,<4.0"

dependencies = [
    "langchain==0.2.1",                 # Pin base version
    "langchain-community",              # Let Poetry choose compatible versions
    "langchain-core",
    "langchain-google-genai",
    "langchain-huggingface",
    "chromadb",
    "python-dotenv",
    "PyMuPDF",                          # For PDF loading
    "tiktoken",                         # Optional tokenizer
    "sentence-transformers",           # If using SentenceTransformer directly
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
